Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/orthogonal-parent-shrink-to-fit-001q-expected.txt


FAIL 17: Shrink-to-fit table-cell with a child of orthogonal block assert_equals: the left edge of the orange box touches the right edge of the blue box expected 0 but got -20
Test passes if the X-position of the left edge of the orange box and the right edge of the blue box are the same.

If script is enabled, there should be one or more PASS and no FAIL.

17: Shrink-to-fit table-cell with a child of orthogonal block

HH
ZZ
